Identity Server Cleanup
